% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_age}
\alias{.get_age}
\title{Calculate age}
\usage{
.get_age(from_date, to_date = lubridate::now(), dec = FALSE)
}
\arguments{
\item{from_date}{from date}

\item{to_date}{to date}

\item{dec}{decimal age}
}
\description{
Calculate age
}
