% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_rosters}
\alias{.get_rosters}
\title{Pull \code{nflreadr} rosters}
\usage{
.get_rosters(seasons)
}
\arguments{
\item{seasons}{choose season or range of seasons}
}
\description{
Pull \code{nflreadr} rosters
}
\keyword{internal}
