% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep_build.R
\name{ep_build}
\alias{ep_build}
\title{Preprocess Data}
\usage{
ep_build(season = nflreadr:::most_recent_season())
}
\arguments{
\item{season}{a numeric vector of seasons that defaults to most recent season. Must be later than 2006.}
}
\value{
a list containing two dataframes: pbp for play-level ep predictions and weekly for game-level summaries.
}
\description{
This function performs the necesssary pre-processing steps to make
expected points predictions on \code{nflreadr} data
}
\examples{
\donttest{
try({ # prevents cran-related errors
  ep_build(season = 2021)
})
}

}
