% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep_version.R
\name{ep_cache_models}
\alias{ep_cache_models}
\title{Model versioning}
\usage{
ep_cache_models(
  version = c("latest", "v1.0.0"),
  force = FALSE,
  ask = interactive()
)
}
\arguments{
\item{version}{one of "latest" or "v1.0.0" - currently these refer to the same thing}

\item{force}{TRUE or FALSE - forces download regardless of currently existing}

\item{ask}{TRUE or FALSE - ask before downloading - \code{force} will skip this.}
}
\value{
a status message after attempting to download the model.
}
\description{
This function checks the cache for a previously downloaded model and then (optionally) tries to download the model from GitHub release.
}
