% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep_calculate_player_stats.R
\name{ep_calculate_player_stats}
\alias{ep_calculate_player_stats}
\title{Summarize EP}
\usage{
ep_calculate_player_stats(
  predicted_pbp,
  stat_type = c("expected_points", "team_stats")
)
}
\value{
a dataframe with the expected fields added
}
\description{
This function summarizes the EP data up to the game level
}
\examples{
\donttest{
pbp <- nflreadr::load_pbp(2021)
temp <- ep_preprocess(pbp)
predicted_pbp <- ep_predict(temp)
cleaned <- ep_calculate_player_stats(predicted_pbp)

predicted_pbp$pass_df \%>\% filter(game_id == "2021_01_ARI_TEN", posteam == "ARI", first_down == 1, penalty == 1) \%>\% tibble::view()
predicted_pbp$rush_df \%>\% filter(week == 5, two_point_attempt == 1) \%>\% view()
}

}
\seealso{
\code{vignette("basic")} for example usage
}
