% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep_load.R
\name{ep_load}
\alias{ep_load}
\title{Load Expected Points data}
\usage{
ep_load(
  season = nflreadr:::most_recent_season(),
  type = c("weekly", "pbp_pass", "pbp_rush"),
  version = c("latest", "v1.0.0")
)
}
\arguments{
\item{season}{A numeric vector of four digit years associated with given NFL seasons - defaults to latest season.}

\item{type}{Data type - one of \code{"weekly"}, \code{"pbp_pass"}, \code{"pbp_rush"}, or \code{"all"}}

\item{version}{EP model version: one of "latest" (default) or "v1.0.0" - these are currently identical.}
}
\value{
a dataframe identical to what would be returned by \code{ffopportunity::ep_build()} for a given season.
}
\description{
This function downloads precomputed expected points data from the ffopportunity automated releases.
}
\examples{
\donttest{
try({
  ep_load() \%>\% head(10)
  ep_load(2020:2021) \%>\% head(10)
  ep_load(2021, type = "pbp_pass") \%>\% head(10)
  ep_load(2006, type = "pbp_rush", version = "v1.0.0") \%>\% head(10)
})
}

}
\seealso{
Other main: 
\code{\link{ep_build}()}
}
\concept{main}
