% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep_load.R
\name{ep_load_pbp_pass}
\alias{ep_load_pbp_pass}
\title{Load EP Pass Play By Play}
\usage{
ep_load_pbp_pass(
  season = nflreadr:::most_recent_season(),
  version = c("latest", "v1.0.0")
)
}
\arguments{
\item{season}{A numeric vector of 4-digit years associated with given NFL seasons - defaults to latest season. If set to \code{TRUE}, returns all available data since 1999.}

\item{version}{EP model version: one of "latest" or "v1.0.0" (these are identical at the moment)}
}
\value{
The complete ffexpectedpoints dataset as returned by \code{ffexpectedpoints::build_nflfastR_pbp()}
(see below) for all given \code{seasons}
}
\description{
Loads multiple seasons from the
\href{https://github.com/ffverse/ffexpectedpoints-data}{ffexpectedpoints data repository}
}
\examples{
\donttest{
  try({ep_load(2019:2020)})
}

}
