% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep_load.R
\name{ep_load_player_stats}
\alias{ep_load_player_stats}
\title{Load Player Level Weekly EP Stats}
\usage{
ep_load_player_stats(
  season = nflreadr:::most_recent_season(),
  version = c("latest", "v1.0.0")
)
}
\arguments{
\item{season}{A numeric vector of 4-digit years associated with given NFL
seasons - defaults to latest season. If set to \code{TRUE},
returns all available data since 1999.}

\item{version}{EP model version: one of "latest" or "v1.0.0" (these are identical at the moment)}
}
\value{
A tibble of week-level player stats with corresponding EP variables
}
\description{
Load Player Level Weekly EP Stats
}
\examples{
\donttest{
  try({ep_load_player_stats()})
}

}
