% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep_predict.R
\name{ep_predict}
\alias{ep_predict}
\title{Predict EP}
\usage{
ep_predict(preprocessed_pbp)
}
\value{
a dataframe with the expected fields added
}
\description{
This function calls the prediction objects on the pre-processed data
}
\examples{
\donttest{
pbp <- nflreadr::load_pbp(2021)
preprocessed_pbp <- ep_preprocess(pbp)
pred_obj <- ep_predict(preprocessed_pbp)
}

}
\seealso{
\code{vignette("basic")} for example usage
}
