% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffopportunity-package.R
\docType{package}
\name{ffopportunity-package}
\alias{ffopportunity}
\alias{ffopportunity-package}
\title{ffopportunity: Models for Fantasy Football Expected Points}
\description{
Downloads expected fantasy points data from 'ffverse' repositories if available, and otherwise builds up expected points data by applying models to 'nflverse' play-by-play data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ffopportunity.ffverse.com}
  \item \url{https://github.com/ffverse/ffopportunity}
  \item Report bugs at \url{https://github.com/ffverse/ffopportunity/issues}
}

}
\author{
\strong{Maintainer}: Joe Sydlowski \email{syd235@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Tan Ho \email{tan@tanho.ca} (\href{https://orcid.org/0000-0001-8388-5155}{ORCID})
}

}
\keyword{internal}
