% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep_load.R
\name{load_ep_pbp_pass}
\alias{load_ep_pbp_pass}
\title{Load EP Pass Play By Play}
\usage{
load_ep_pbp_pass(season = nflreadr:::most_recent_season(), file_type = "rds")

load_ep_pbp_pass(season = nflreadr:::most_recent_season(), file_type = "rds")
}
\arguments{
\item{season}{A numeric vector of 4-digit years associated with given NFL seasons - defaults to latest season. If set to \code{TRUE}, returns all available data since 1999.}

\item{file_type}{One of \code{"rds"}, \code{"parquet"}, or \code{"csv"}.}
}
\value{
The complete ffexpectedpoints dataset as returned by \code{ffexpectedpoints::build_nflfastR_pbp()}
(see below) for all given \code{seasons}

The complete ffexpectedpoints dataset as returned by \code{ffexpectedpoints::build_nflfastR_pbp()}
(see below) for all given \code{seasons}
}
\description{
Loads multiple seasons from the
\href{https://github.com/ffverse/ffexpectedpoints-data}{ffexpectedpoints data repository}

Loads multiple seasons from the
\href{https://github.com/ffverse/ffexpectedpoints-data}{ffexpectedpoints data repository}
}
\examples{
\donttest{
  try({load_ep_pbp_pass(2019:2020)})
}

\donttest{
  try({load_ep_pbp_rush(2019:2020)})
}

}
