% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep_load.R
\name{load_player_ep_stats}
\alias{load_player_ep_stats}
\title{Load Player Level Weekly EP Stats}
\usage{
load_player_ep_stats(
  season = nflreadr:::most_recent_season(),
  file_type = "rds"
)
}
\arguments{
\item{season}{A numeric vector of 4-digit years associated with given NFL
seasons - defaults to latest season. If set to \code{TRUE},
returns all available data since 1999.}

\item{file_type}{One of \code{"rds"}, \code{"parquet"}, or \code{"csv"}.}
}
\value{
A tibble of week-level player stats with corresponding EP variables
}
\description{
Load Player Level Weekly EP Stats
}
\examples{
\donttest{
  try({load_player_ep_stats()})
}

}
