% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep_calculate_player_stats.R
\name{ep_calculate_player_stats}
\alias{ep_calculate_player_stats}
\title{Summarize EP}
\usage{
ep_calculate_player_stats(predicted_pbp)
}
\value{
a dataframe with the expected fields added
}
\description{
This function summarizes the EP data up to the game level
}
\examples{
\donttest{
pbp <- nflreadr::load_pbp(2021)
temp <- ep_preprocess(pbp)
pred_obj <- ep_predict(temp)
cleaned <- ep_calculate_player_stats(pred_obj)
}

}
\seealso{
\code{vignette("basic")} for example usage
}
