% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep_predict.R
\name{ep_predict}
\alias{ep_predict}
\title{Predict EP}
\usage{
ep_predict(preprocessed_pbp, version = c("latest", "v1.0.0"))
}
\arguments{
\item{preprocessed_pbp}{list with dataframes created by \code{ep_preprocess}}

\item{version}{ep model version - available is \code{"latest"} or \code{"v1.0.0"} (these are currently the same thing)}
}
\value{
a dataframe with the expected fields added
}
\description{
This function runs the prediction functions over preprocessed data.
}
\examples{
\donttest{
  try({
  preprocessed <- readRDS(system.file("ep_preprocessed.rds",package = "ffopportunity"))
  # this file is equivalent to nflreadr::load_pbp(2021) \%>\% head(1000) \%>\% ep_preprocess()
  ep_predict(preprocessed)
  })
}

}
