% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep_preprocess.R
\name{ep_preprocess}
\alias{ep_preprocess}
\title{Preprocess Data}
\usage{
ep_preprocess(pbp)
}
\value{
a list of two dataframes (one for passes and one for rushes)
of nflreadr data with the expectedpoints columns transformed for prediction
}
\description{
This function performs the necesssary pre-processing steps to make
expected points predictions on \code{nflreadr} data
}
\examples{
\donttest{
pbp <- nflreadr::load_pbp(2021)
.preprocess_common_fields(pbp) \%>\% dplyr::glimpse()
temp <- ep_preprocess(pbp) \%>\% dplyr::glimpse()
}

}
\seealso{
\code{vignette("basic")} for example usage
}
