% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep_preprocess.R
\name{ep_preprocess}
\alias{ep_preprocess}
\title{Preprocess Data}
\usage{
ep_preprocess(pbp)
}
\arguments{
\item{pbp}{pbp dataframe from \code{nflreadr::load_pbp()}}
}
\value{
a list of two dataframes (one for passes and one for rushes)
of nflreadr data with the expectedpoints columns transformed for prediction
}
\description{
This function performs pre-processing steps to make
expected points predictions on \code{nflreadr} data
}
\examples{
\donttest{
try({ # catch failures for CRAN purposes
  preprocessed <- readRDS(system.file("pbp_download.rds",package = "ffopportunity"))
  # this file is equivalent to nflreadr::load_pbp(2021) \%>\% head(100)
  ep_preprocess(preprocessed)
  })
}

}
\seealso{
\code{vignette("basic")} for example usage
}
