% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep_summarize.R
\name{ep_summarize}
\alias{ep_summarize}
\title{Summarize EP}
\usage{
ep_summarize(
  predicted_pbp,
  stat_type = c("all", "expected_points", "team_stats")
)
}
\arguments{
\item{predicted_pbp}{list with dataframes created by \code{ep_predict}}

\item{stat_type}{option to limit the columns returned by \code{ep_summarize}}
}
\value{
a dataframe with the expected fields added
}
\description{
This function summarizes the EP data up to the game level
}
\examples{
\donttest{
try({
nflreadr::load_pbp(2021) \%>\%
  head(1000) \%>\%
  ep_preprocess() \%>\%
  ep_predict() \%>\%
  ep_summarize()
})
}

}
\seealso{
\code{vignette("basic")} for example usage
}
