% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep_summarize.R
\name{ep_summarize}
\alias{ep_summarize}
\title{Summarize EP}
\usage{
ep_summarize(predicted_pbp, stat_type = c("expected_points", "team_stats"))
}
\value{
a dataframe with the expected fields added
}
\description{
This function summarizes the EP data up to the game level
}
\examples{
\donttest{
pbp <- nflreadr::load_pbp(2021)
temp <- ep_preprocess(pbp)
predicted_pbp <- ep_predict(temp)
cleaned <- ep_summarize(predicted_pbp)

predicted_pbp$pass_df \%>\%
  dplyr::filter(game_id == "2021_01_ARI_TEN", posteam == "ARI", first_down == 1, penalty == 1)
predicted_pbp$rush_df \%>\% dplyr::filter(week == 5, two_point_attempt == 1)
}

}
\seealso{
\code{vignette("basic")} for example usage
}
